pyproject.toml
[project]
name = "yojimbo"
version = "0.1.0"
description = "Real-Time Multi-Lingual AI Agent Receptionist for Local Government"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "twilio>=9.0.0",
    "google-genai>=1.0.0",
    "google-cloud-translate>=3.15.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "itsdangerous>=2.1.0",
    "starlette-csrf>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
